# 12 optimize queries of account app