# Report

## accounts
- #### [profile/followings/](http://localhost:8000/api/accounts/profile/followings/) 
    - Before optimization: 12 queries in 9.16ms
    - After optimization: 3 queries in 3.82ms
    - **Result: 75% optimization in queries and 58% optimization in time**


- #### [profile/followers/](http://localhost:8000/api/accounts/profile/followers/) 
    - Before optimization: 12 queries in 8.95ms
    - After optimization: 3 queries in 4.58ms
    - **Result: 75% optimization in queries and 49% optimization in time**

## jobs
- #### [userskillprojects/](http://localhost:8000/api/jobs/userskillprojects/) 
    - Before optimization: 30 queries in 35.91ms 
    - After optimization: 24 queries in 20.28ms
    - **Result: 20% optimization in queries and 43% optimization in time**